Hello my sisters, I want to teach programming in hacking and penetration
Sisters, I have been betrayed several times and I am tired of this situation. I am looking for a program in which the victim’s mobile phone was hacked by sending the link and I have not found it until today. Does anyone of you know what the name of the program is? I want to get married and now